requirements.txt
Django
Pillow
dialogflow